test: keep remote URL regression coverage scoped
